TRAD-13 Implement service to get onboardDate from market_info by symbol

+++ src/main/java/com/tradey/technical_analysis/domain/entity/SymbolInfoEntity.java
@@ -3,12 +3,10 @@ package com.tradey.technical_analysis.domain.entity;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
-import java.math.BigInteger;
-
 @RequiredArgsConstructor
 public class SymbolInfoEntity {
     private final String exchangeType;
     private final String symbol;
     @Getter
-    private final BigInteger onboardDate;
+    private final long onboardDate;
 }
